+++ client/src/components/recommend/ResultBoard.vue
